requirements.txt
aiohttp>=3.11
aiosqlite>=0.21
python-dotenv>=1.1
